- Code refactoring for building 3DSX and running under Homebrew GridLauncher - Optimisations to Mode 7 (modest improvement gains for F-Zero) - Added options to force frame rate to 50 FPS, 60 FPS, or based on ROM’s region - Added options to allow screen stretching to 4:3 (320x240) - Captures top screen screenshot in 400 x 240 only - Minor optimisations to main CPU emulation. - Reduce the SRAM auto-save to 100 milliseconds (approx 5-6 frames after the last SRAM write) - Handles Mode 7 tile 0 repeat (allows Super Mario Kart to display area outside of the roads correctly) - Fixed SRAM reading bug that causes DKC1 trigger piracy screen.

+++ source/snes9x.h
@@ -216,7 +216,7 @@ struct SICPU
     uint8  *Speed;
     struct SOpcodes *S9xOpcodes;
     uint8  _Carry;
-    uint8  _Zero;
+    uint16  _Zero;
     uint8  _Negative;
     uint8  _Overflow;
     bool8  CPUExecuting;
